Fix bug that enemy hp is showing incorrectly and rare bug in one of the combos

+++ RhythmTest/Assets/Scripts/StoneManager.cs
@@ -76,7 +76,7 @@ public class StoneManager : MonoBehaviour {
 
     private void UpdateEnemyHealth()
     {
-        int enemyCurrHealth = (int) Mathf.Ceil((float) (enemyRef.health / 10));
+        int enemyCurrHealth = (int) Mathf.Ceil(enemyRef.health / 10f);
 
         for (int i = 0; i < enemyCurrHealth; i++)
         {
